requirements.txt
# === Core Framework ===
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
bcrypt==4.0.1
pydantic==2.10.4
email-validator==2.2.0

# === Base de Datos ===
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# === Inteligencia Artificial y ML ===
groq==0.14.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.3.2
scikit-fuzzy==0.5.0
networkx==3.2.1
tensorflow==2.16.1
keras==3.3.3
spacy==3.7.2
joblib==1.3.2

# === Modelos de spaCy ===
es_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.7.0/es_core_news_sm-3.7.0-py3-none-any.whl

# === Integraciones Externas ===
firebase-admin==6.7.0
resend==2.6.0

# === Utilidades ===
requests==2.32.3
jinja2==3.1.5
websockets==14.1